Control.Concurrent.Find

Description

Concurrent nondeterministic search.

Find monad

data Find m a

Running searches

runFind :: MonadConc m => Find m a -> m (Maybe a)

Combinators

(!) :: MonadConc m => [a] -> (a -> Bool) -> Find m a

Flipped infix version of findIn.

success :: MonadConc m => a -> Find m a

Search which always succeeds.

failure :: MonadConc m => Find m a

Search which always fails.

findIn :: MonadConc m => (a -> Bool) -> [a] -> Find m a

Find an element of a list satisfying a predicate.

findBoth :: MonadConc m => (a -> Bool) -> (b -> Bool) -> [a] -> [b] -> Find m (a, b)

Find elements from a pair of lists satisfying predicates. Both lists are searched in parallel.

findEither :: MonadConc m => (a -> Bool) -> (b -> Bool) -> [a] -> [b] -> Find m (Either a b)

Find an element from one of two lists which satisfies a predicate. Both lists are searched in parallel.

oneOf :: MonadConc m => [Find m a] -> Find m a

Return one non-failing result nondeterministically.