Control.Concurrent.Find

Description

Concurrent nondeterministic search.

Find monad

data Find m a

A value of type Find m a represents a concurrent search computation (happening in the MonadConc monad m) which may produce a value of type a, or fail. If a value can be returned, one will be (although it's nondeterministic which one will actually be returned). Usually you will be working with values of type Find IO a, but the generality allows for testing.

You should prefer using the Applicative instance over the Monad instance if you can, as the Applicative preserves parallelism.

runFind :: MonadConc m => Find m a -> m (Maybe a)

Execute a Find computation, returning a Just value if there was at least one result (and a different value may be returned each time), or Nothing if there are no results.

unsafeRunFind :: MonadConc m => Find m a -> m a

Unsafe version of runFind. This will error at runtime if the computation fails.

hasResult :: Find IO a -> Bool

Check if a computation has a result. This will block until a result is found.

Combinators

(!) :: MonadConc m => [a] -> (a -> Bool) -> Find m a

Flipped infix version of findIn.

success :: MonadConc m => a -> Find m a

Search which always succeeds.

failure :: MonadConc m => Find m a

Search which always fails.

findIn :: MonadConc m => (a -> Bool) -> [a] -> Find m a

Find an element of a list satisfying a predicate.

findBoth :: MonadConc m => (a -> Bool) -> (b -> Bool) -> [a] -> [b] -> Find m (a, b)

Find elements from a pair of lists satisfying predicates. Both lists are searched in parallel.

findEither :: MonadConc m => (a -> Bool) -> (b -> Bool) -> [a] -> [b] -> Find m (Either a b)

Find an element from one of two lists which satisfies a predicate. Both lists are searched in parallel.

oneOf :: MonadConc m => [Find m a] -> Find m a

Return one non-failing result nondeterministically.