Control.Concurrent.Find

Description

Concurrent nondeterministic search.

Find monad

data Find m a

A value of type Find m a represents a concurrent search computation (happening in the MonadConc monad m) which may produce a value of type a, or fail. If a value can be returned, one will be (although it's nondeterministic which one will actually be returned). Usually you will be working with values of type Find IO a, but the generality allows for testing.

You should prefer using the Applicative instance over the Monad instance if you can, as the Applicative preserves parallelism.

runFind :: MonadConc m => Find m a -> m (Maybe a)

Execute a Find computation, returning a Just value if there was at least one result (and a different value may be returned each time), or Nothing if there are no results.

unsafeRunFind :: MonadConc m => Find m a -> m a

Unsafe version of runFind. This will error if the computation fails.

hasResult :: Find IO a -> Bool

Check if a computation has a result. This will block until a result is found.

toMaybe :: Find IO a -> Maybe a

Convert a Find to a Maybe. This is only safe if the result found is deterministic (eg, generated by firstOf).

Example safe usage:

first :: (a -> Bool) -> [a] -> Maybe a
first p xs = toMaybe $ xs .! p

This uses unsafePerformIO internally.

Result Streams

data Stream m a

A value of type Stream m a represents a streaming sequence of results being computed concurrently in the MonadConc monad m, which contains some values of type a.

readStream :: MonadConc m => Stream m a -> m (Maybe a)

Read the head of stream, if the stream is finished Nothing will be returned.

takeStream :: MonadConc m => Int -> Stream m a -> m [a]

Take some values from the start of a stream, if the stream does not contain that many elements, a shorter result list is returned.

unsafeReadStream :: MonadConc m => Stream m a -> m a

Unsafe version of readStream. This will error if the stream is empty.

gatherStream :: (MonadConc m, Monoid o) => Stream m o -> m o

Gather all the values of the stream together. This will block until the entire computation terminates.

zipStream :: MonadConc m => (a -> b -> c) -> Maybe a -> Maybe b -> Stream m a -> Stream m b -> Stream m c

Zip two streams together, with an optional values to supply when a component stream runs out. If no value is given, the resulting stream ends as soon as that component stream does.

toList :: Stream IO a -> [a]

Convert a Stream to a list. This is only safe if the stream is never used again after calling this function, and the order of the stream is deterministic (eg, generated by orderedOf).

Example safe usage:

filterPar :: (a -> Bool) -> [a] -> [a]
filterPar p xs = toList . unsafePerformIO $ xs >! p

This uses unsafePerformIO internally.

Basic Searches

success :: MonadConc m => a -> Find m a

Search which always succeeds.

failure :: MonadConc m => Find m a

Search which always fails.

oneOf :: MonadConc m => [Find m a] -> Find m a

Return one non-failing result nondeterministically.

allOf :: MonadConc m => [Find m a] -> m (Stream m a)

Return all non-failing results, the order is nondeterministic.

firstOf :: MonadConc m => [Find m a] -> Find m a

Return the first non-failing result. This is deterministic.

orderedOf :: MonadConc m => [Find m a] -> m (Stream m a)

Return all non-failing results, in the order they appear in the list. This is deterministic.

Combinators

(!) :: MonadConc m => [a] -> (a -> Bool) -> Find m a

Flipped infix version of find.

(?) :: MonadConc m => [a] -> (a -> Maybe b) -> Find m b

Flipped infix version of mapped.

(@!) :: MonadConc m => [a] -> (a -> Bool) -> m (Stream m a)

Flipped infix version of findAll.

(@?) :: MonadConc m => [a] -> (a -> Maybe b) -> m (Stream m b)

Flipped infix version of mappedAll.

find :: MonadConc m => (a -> Bool) -> [a] -> Find m a

Return an element in the list matching the predicate.

findAll :: MonadConc m => (a -> Bool) -> [a] -> m (Stream m a)

Variant of find which returns all such elements.

mapped :: MonadConc m => (a -> Maybe b) -> [a] -> Find m b

Return an element in the list giving a Just.

mappedAll :: MonadConc m => (a -> Maybe b) -> [a] -> m (Stream m b)

Variant of mapped which returns all such elements.

both :: MonadConc m => (a -> Bool) -> (b -> Bool) -> [a] -> [b] -> Find m (a, b)

Find elements from a pair of lists satisfying predicates. Both lists are searched in parallel.

either :: MonadConc m => (a -> Bool) -> (b -> Bool) -> [a] -> [b] -> Find m (Either a b)

Find an element from one of two lists which satisfies a predicate. Both lists are searched in parallel.

merging :: (MonadConc m, Monoid o, Eq o) => (a -> o) -> [a] -> Find m o

Gather all non-failing results and mconcat them together. The order of merging is nondeterministic, and so this is most likely to be useful in the context of gathering a list of results.

If there are likely to be many mempty results, this is faster than

fmap mconcat $ xs @? Just . f

Deterministic

These are all safe to extract the result from with toMaybe and toList.

(.!) :: MonadConc m => [a] -> (a -> Bool) -> Find m a

Flipped infix version of first.

(.?) :: MonadConc m => [a] -> (a -> Maybe b) -> Find m b

Flipped infix version of mappedFirst.

(>!) :: MonadConc m => [a] -> (a -> Bool) -> m (Stream m a)

Flipped infix version of inOrder.

(>?) :: MonadConc m => [a] -> (a -> Maybe b) -> m (Stream m b)

Flipped infix version of mappedInOrder.

first :: MonadConc m => (a -> Bool) -> [a] -> Find m a

Variant of find which returns the first such element.

inOrder :: MonadConc m => (a -> Bool) -> [a] -> m (Stream m a)

Variant of findAll which returns the elements in the same order that they appear in the original list.

mappedFirst :: MonadConc m => (a -> Maybe b) -> [a] -> Find m b

Variant of mapped which returns the first such element.

mappedInOrder :: MonadConc m => (a -> Maybe b) -> [a] -> m (Stream m b)

Variant of mappedAll which returns the elements in the same order that they appear in the original list.